requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.9.0
numpy>=1.26.0
ollama>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
